mergesort files
"mergesort_parallel_pool&map_np.py" is the main file. the other two, "mergesort_sequential_np", and its other variation (with number "2" at the end), are the attempts i did to convert them to use np, but it seems that running the main file with the first version of the mergesort seems to give results that I were not expecting (using numpy, sequential is still faster than parallel, as opposed to not using numpy), hence it's still not done and have to go over/study them.